[refactor] rename strategy
