--- hummingbot/strategy/base_price_action/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-
-from .base_price_action import BasePriceActionStrategy
-
-
-__all__ = [
-    BasePriceActionStrategy
-]
