+++ hummingbot/strategy/self_trade/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+from .self_trade import SelfTradeStrategy
+
+
+__all__ = [
+    SelfTradeStrategy
+]
