+++ hummingbot/strategy/self_trade/trade_band.py
